Update module name to better reflect its purpose
