--- embabel-agent-autoconfigure/models/embabel-agent-openai-compatible-autoconfigure/pom.xml
@@ -1,46 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 https://maven.apache.org/xsd/maven-4.0.0.xsd">
-    <modelVersion>4.0.0</modelVersion>
-    <parent>
-        <groupId>com.embabel.agent</groupId>
-        <artifactId>embabel-agent-autoconfigure</artifactId>
-        <version>0.3.2-SNAPSHOT</version>
-        <relativePath>../../pom.xml</relativePath>
-    </parent>
-    <artifactId>embabel-agent-openai-compatible-autoconfigure</artifactId>
-    <packaging>jar</packaging>
-    <name>Embabel Agent Autoconfiguration Models OpenAI Compatible</name>
-    <description>OpenAI Compatible models Auto-Configuration for Embabel Agent API</description>
-    <url>https://github.com/embabel/embabel-agent</url>
-
-    <scm>
-        <url>https://github.com/embabel/embabel-agent</url>
-        <connection>scm:git:https://github.com/embabel/embabel-agent.git</connection>
-        <developerConnection>scm:git:https://github.com/embabel/embabel-agent.git</developerConnection>
-        <tag>HEAD</tag>
-    </scm>
-
-    <dependencies>
-        <dependency>
-            <groupId>com.embabel.agent</groupId>
-            <artifactId>embabel-agent-api</artifactId>
-        </dependency>
-
-        <dependency>
-            <groupId>com.embabel.agent</groupId>
-            <artifactId>embabel-agent-openai</artifactId>
-        </dependency>
-
-        <dependency>
-            <groupId>org.springframework.ai</groupId>
-            <artifactId>spring-ai-openai</artifactId>
-        </dependency>
-
-        <dependency>
-            <groupId>com.embabel.agent</groupId>
-            <artifactId>embabel-agent-test-internal</artifactId>
-            <scope>test</scope>
-        </dependency>
-    </dependencies>
-
-</project>
