--- embabel-agent-autoconfigure/models/embabel-agent-openai-compatible-autoconfigure/src/main/kotlin/com/embabel/agent/config/models/openai/compatible/OpenAiCompatibleModelsConfig.kt
@@ -1,165 +1,0 @@
-/*
- * Copyright 2024-2025 Embabel Software, Inc.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package com.embabel.agent.config.models.openai.compatible
-
-import com.embabel.agent.openai.OpenAiCompatibleModelFactory
-import com.embabel.agent.openai.StandardOpenAiOptionsConverter
-import com.embabel.agent.spi.common.RetryProperties
-import com.embabel.common.ai.autoconfig.ProviderInitialization
-import com.embabel.common.ai.autoconfig.RegisteredModel
-import com.embabel.common.ai.model.Llm
-import com.embabel.common.util.ExcludeFromJacocoGeneratedReport
-import io.micrometer.observation.ObservationRegistry
-import org.springframework.beans.factory.ObjectProvider
-import org.springframework.beans.factory.annotation.Value
-import org.springframework.beans.factory.config.ConfigurableBeanFactory
-import org.springframework.boot.context.properties.ConfigurationProperties
-import org.springframework.boot.context.properties.EnableConfigurationProperties
-import org.springframework.context.annotation.Bean
-import org.springframework.context.annotation.Configuration
-
-/**
- * Configuration properties for OpenAI Compatible model settings.
- * These properties can be set in application.properties/yaml using the
- * prefix embabel.agent.platform.models.openai.compatible
- */
-@ConfigurationProperties(prefix = "embabel.agent.platform.models.openai.compatible")
-class OpenAiCompatibleProperties : RetryProperties {
-    /**
-     *  Maximum number of attempts.
-     */
-    override var maxAttempts: Int = 10
-
-    /**
-     * Initial backoff interval (in milliseconds).
-     */
-    override var backoffMillis: Long = 5000L
-
-    /**
-     * Backoff interval multiplier.
-     */
-    override var backoffMultiplier: Double = 5.0
-
-    /**
-     * Maximum backoff interval (in milliseconds).
-     */
-    override var backoffMaxInterval: Long = 180000L
-}
-
-/**
- * Configuration for OpenAI Compatible language and embedding models.
- * You can specify custom models via the
- * `OPENAI_CUSTOM_MODELS` environment variable (comma-separated list). This is useful
- * when using OpenAI-compatible APIs (like Groq, Together AI, etc.) that may use
- * different model names.
- *
- * Example:
- * ```
- * OPENAI_COMPATIBLE_BASE_URL=https://api.groq.com/openai/v1
- * OPENAI_COMPATIBLE_API_KEY=your-api-key
- * OPENAI_COMPATIBLE_CUSTOM_MODELS=llama-3.3-70b-versatile,mixtral-8x7b-32768,gemma2-9b-it
- * EMBABEL_MODELS_DEFAULT_LLM=llama-3.3-70b-versatile
- * ```
- *
- * All custom models will be registered and available. Use `EMBABEL_MODELS_DEFAULT_LLM`
- * to specify which model should be the default.
- */
-@Configuration(proxyBeanMethods = false)
-@EnableConfigurationProperties(OpenAiCompatibleProperties::class)
-@ExcludeFromJacocoGeneratedReport(reason = "OpenAi Compatible configuration can't be unit tested")
-class OpenAiCompatibleModelsConfig(
-    @Value("\${OPENAI_COMPATIBLE_BASE_URL:#{null}}")
-    baseUrl: String?,
-    @Value("\${OPENAI_COMPATIBLE_API_KEY}")
-    apiKey: String,
-    @Value("\${OPENAI_COMPATIBLE_CUSTOM_MODELS:#{null}}")
-    private val customModels: String?,
-    observationRegistry: ObjectProvider<ObservationRegistry>,
-    private val properties: OpenAiCompatibleProperties,
-    private val configurableBeanFactory: ConfigurableBeanFactory,
-) : OpenAiCompatibleModelFactory(
-    baseUrl = baseUrl,
-    apiKey = apiKey,
-    completionsPath = null,
-    embeddingsPath = null,
-    observationRegistry = observationRegistry.getIfUnique { ObservationRegistry.NOOP }
-) {
-
-    private val customModelList: List<String> = customModels
-        ?.split(",")
-        ?.map { it.trim() }
-        ?.filter { it.isNotBlank() }
-        ?: emptyList()
-
-    init {
-        logger.info("OpenAI Compatible models are available: {}", properties)
-        if (customModelList.isNotEmpty()) {
-            logger.info("Custom OpenAI-compatible models configured: {}", customModelList)
-        }
-    }
-
-    @Bean
-    fun openAiCompatibleModelsInitializer(): ProviderInitialization {
-        val registeredLlms = registerCustomModels()
-
-        return ProviderInitialization(
-            provider = CUSTOM_PROVIDER,
-            registeredLlms = registeredLlms,
-        ).also { logger.info(it.summary()) }
-    }
-
-    /**
-     * Registers custom models specified via OPENAI_CUSTOM_MODELS environment variable.
-     */
-    private fun registerCustomModels(): List<RegisteredModel> {
-        return customModelList.map { modelId ->
-            try {
-                val llm = createCustomLlm(modelId)
-                configurableBeanFactory.registerSingleton(modelId, llm)
-                logger.info(
-                    "Registered custom OpenAI-compatible model bean: {}",
-                    modelId
-                )
-                RegisteredModel(beanName = modelId, modelId = modelId)
-            } catch (e: Exception) {
-                logger.error("Failed to create custom model: {}", modelId, e)
-                throw e
-            }
-        }
-    }
-
-    /**
-     * Creates an LLM for a custom model specified via OPENAI_CUSTOM_MODELS.
-     * Uses standard OpenAI options converter since we don't know the model's capabilities.
-     */
-    private fun createCustomLlm(modelId: String): Llm {
-        val chatModel = chatModelOf(
-            model = modelId,
-            retryTemplate = properties.retryTemplate(modelId)
-        )
-
-        return Llm(
-            name = modelId,
-            model = chatModel,
-            provider = CUSTOM_PROVIDER,
-            optionsConverter = StandardOpenAiOptionsConverter,
-        )
-    }
-
-    companion object {
-        private const val CUSTOM_PROVIDER = "OpenAI-Compatible"
-    }
-}
