+++ embabel-agent-autoconfigure/models/embabel-agent-openai-custom-autoconfigure/src/main/java/com/embabel/agent/autoconfigure/models/openai/custom/AgentOpenAiCustomAutoConfiguration.java
@@ -1,0 +1,37 @@
+/*
+ * Copyright 2024-2025 Embabel Software, Inc.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package com.embabel.agent.autoconfigure.models.openai.custom;
+
+import com.embabel.agent.config.models.openai.custom.OpenAiCustomModelsConfig;
+import org.springframework.boot.autoconfigure.AutoConfiguration;
+import org.springframework.boot.autoconfigure.AutoConfigureBefore;
+import org.springframework.context.annotation.Import;
+
+/**
+ * Autoconfiguration for OpenAI Custom models in the Embabel Agent system.
+ * <p>
+ * This class serves as a Spring Boot autoconfiguration entry point that:
+ * - Scans for configuration properties in the "com.embabel.agent" package
+ * - Imports the {@link OpenAiCompatibleModelsConfig} configuration to register OpenAI model beans
+ * <p>
+ * The configuration is automatically activated when the OpenAI models
+ * dependencies are present on the classpath.
+ */
+@AutoConfiguration
+@AutoConfigureBefore(name = {"com.embabel.agent.autoconfigure.platform.AgentPlatformAutoConfiguration"})
+@Import(OpenAiCustomModelsConfig.class)
+public class AgentOpenAiCustomAutoConfiguration {
+}
