+++ embabel-agent-dependencies/pom.xml
@@ -110,7 +110,7 @@
 
             <dependency>
                 <groupId>com.embabel.agent</groupId>
-                <artifactId>embabel-agent-openai-compatible-autoconfigure</artifactId>
+                <artifactId>embabel-agent-openai-custom-autoconfigure</artifactId>
                 <version>${project.version}</version>
             </dependency>
