--- embabel-agent-starters/embabel-agent-starter-openai-compatible/pom.xml
@@ -1,33 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 https://maven.apache.org/xsd/maven-4.0.0.xsd">
-    <modelVersion>4.0.0</modelVersion>
-    <parent>
-        <groupId>com.embabel.agent</groupId>
-        <artifactId>embabel-agent-starters</artifactId>
-        <version>0.3.2-SNAPSHOT</version>
-    </parent>
-    <artifactId>embabel-agent-starter-openai-compatible</artifactId>
-    <packaging>jar</packaging>
-    <name>Embabel Agent OpenAI Compatible Starter</name>
-    <description>Embabel Agent OpenAI Compatible Starter</description>
-    <url>https://github.com/embabel/embabel-agent</url>
-
-    <scm>
-        <url>https://github.com/embabel/embabel-agent</url>
-        <connection>scm:git:https://github.com/embabel/embabel-agent.git</connection>
-        <developerConnection>scm:git:https://github.com/embabel/embabel-agent.git</developerConnection>
-        <tag>HEAD</tag>
-    </scm>
-
-    <dependencies>
-        <dependency>
-            <groupId>com.embabel.agent</groupId>
-            <artifactId>embabel-agent-platform-autoconfigure</artifactId>
-        </dependency>
-        <dependency>
-            <groupId>com.embabel.agent</groupId>
-            <artifactId>embabel-agent-openai-compatible-autoconfigure</artifactId>
-        </dependency>
-    </dependencies>
-
-</project>
